requirements.txt
mwclient
requests
PyYAML
pycryptodome
python-dotenv